crex-live-cricket-scraper/requirements.txt
beautifulsoup4
selenium
webdriver_manager
pymongo
requests
apscheduler
python-dotenv